+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog][keepachangelog], and this project adhe
 ### Added
 
 - Support for quick fixes: [#122]
+- Minimum version check for external AGLint installations: [#123]
 
 ## 1.1.6 - 2024-01-04
 
@@ -364,6 +365,7 @@ The format is based on [Keep a Changelog][keepachangelog], and this project adhe
 [#115]: https://github.com/AdguardTeam/VscodeAdblockSyntax/issues/115
 [#119]: https://github.com/AdguardTeam/VscodeAdblockSyntax/issues/119
 [#122]: https://github.com/AdguardTeam/VscodeAdblockSyntax/issues/122
+[#123]: https://github.com/AdguardTeam/VscodeAdblockSyntax/issues/123
 [AGLint]: https://github.com/AdguardTeam/AGLint/blob/master/CHANGELOG.md
 [keepachangelog]: https://keepachangelog.com/en/1.0.0/
 [semver]: https://semver.org/spec/v2.0.0.html